Fix MISRAC for the folder utils
 - MISRA.CT.UNIQUE.ID
  - A tag name shall be a unique identifier
  - Fixed by renaming the tag names

 - MISRA.ETYPE.CATEGORY.DIFFERENT.2012
  - Both operands of an operator in which the
    usual arithmetic conversions are performed
    shall have the same essential typecategory
  - Fixed by typecasting to the required datatype

 - MISRA.FUNC.NOPROT.DEF.2012
  - A compatible declaration shall be visible when
    an object or function with external linkage is
    defined
  - Fixed ny including the required header file

 - MISRA.ETYPE.COMP.CAST.EXPL.DIFFERENT.2012
  - The value of a composite expression shall not
    be cast to a different essential type category
    or a wider essential type
  - Fixed by typecasting to the required datatype

 - MISRA.FUNC.MODIFIEDPAR.2012
  - A function parameter should not be modified
  - Fixed by having a new variable for the function
    parameter which is modified in the function

 - MISRA.STMT.COND.NOT_BOOLEAN.2012
  - The controlling expression of an if statement and
    the controlling expression of an iteration-statement
    shall have essentially Boolean type
  - Fixed by having a boolean expression in the condition

 - MISRA.ETYPE.INAPPR.OPERAND.UNOP.2012
  - Operands shall not be of an inappropriate essential type
  - Fixed by having a boolean expression in the condition

 - MISRA.INCR_DECR.SIDEEFF.2012
  - A full expression containing an increment or decrement
    operator should have no other potential side effects other
    than that caused by the increment or decrement operator
  - Fixed by having a new variable for the function parameter
    which is modified in the function

 - MISRA.ETYPE.INAPPR.OPERAND.INDEXPR.2012
  - Operands shall not be of an inappropriate essential type
  - Fixed by having the required datatype

 - MISRA.BREAK_OR_GOTO.MULTIPLE.2012
  - There should be no more than one break or goto statement
    used to terminate any iteration statement
  - Fixed by using a new flag variable

- MISRA.RETURN.NOT_LAST
  - A function should have a single point of exit at the end
  - Fixed by a new variable for return values/expressions

+++ utils/ioremap.c
@@ -66,12 +66,12 @@ u32 ioremap_internal(u32 a)
 	 */
 	if (((a >= 0x100000U) && (a < 0x44000000U)) ||
 	    ((a >= 0x46000000U) && (a < 0x60000000U))) {
-		addr += CONFIG_ADDR_REMAP_OFFSET;
+		addr += (u32) CONFIG_ADDR_REMAP_OFFSET;
 	}
 	ret = rat_get_free_map_addr_range_user(&lowAddr, &highAddr);
 	if (ret == SUCCESS) {
 		if ((a >= lowAddr) && (a < highAddr)) {
-			addr += CONFIG_ADDR_REMAP_OFFSET;
+			addr += (u32) CONFIG_ADDR_REMAP_OFFSET;
 		}
 	}
 	return addr;
